# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ASPICE Documentation Platform API",
    "description": "Multi-tenant, role-based SaaS API for automating automotive ASPICE / ISO 26262 documentation from C++ source code repositories.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/projects/{project_id}/compare": {
      "get": {
        "tags": ["compare"],
        "summary": "Compare two versions or commits",
        "operationId": "compare",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "name": "current", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Version tag, commit SHA, or version ID to treat as current" },
          { "name": "baseline", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Version tag, commit SHA, or version ID to treat as baseline" }
        ],
        "responses": {
          "200": {
            "description": "Comparison result summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CompareResult" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/compare/documents": {
      "get": {
        "tags": ["compare"],
        "summary": "List changed documents between two versions",
        "operationId": "compareDocuments",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "name": "current", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "baseline", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of changed documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/DocumentDiffView" }
                    },
                    "summary": {
                      "type": "object",
                      "additionalProperties": { "type": "integer" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/projects/{project_id}/compare/documents/{doc_id}": {
      "get": {
        "tags": ["compare"],
        "summary": "Get section-level diff for a document",
        "operationId": "compareDocumentDetail",
        "parameters": [
          { "$ref": "#/components/parameters/ProjectId" },
          { "$ref": "#/components/parameters/DocId" },
          { "name": "current", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "baseline", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Section-level diff",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "document_name": { "type": "string" },
                    "sections": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/SectionDiffView" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Project identifier"
      },
...
      "DocId": {
        "name": "doc_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Document identifier"
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "NOT_FOUND" },
              "message": { "type": "string" },
              "status": { "type": "integer", "example": 404 }
            }
          }
        }
      },
...
      "CompareResult": {
        "type": "object",
        "properties": {
          "current": {
            "type": "object",
            "properties": {
              "ref": { "type": "string" },
              "version": { "type": "string", "nullable": true },
              "branch": { "type": "string" }
            }
          },
          "baseline": {
            "type": "object",
            "properties": {
              "ref": { "type": "string" },
              "version": { "type": "string", "nullable": true },
              "branch": { "type": "string" }
            }
          },
          "summary": { "type": "object", "additionalProperties": { "type": "integer" } },
          "changed_documents": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DocumentDiffView" }
          }
        }
      },
      "DocumentDiffView": {
        "type": "object",
        "properties": {
          "document_id": { "type": "string" },
          "name": { "type": "string" },
          "process": { "type": "string" },
          "diff_type": { "type": "string", "enum": ["added", "changed", "removed", "unchanged"] },
          "sections_changed": { "type": "array", "items": { "type": "string" } }
        }
      },
      "SectionDiffView": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "title": { "type": "string" },
          "diff_type": { "type": "string", "enum": ["changed", "unchanged"] },
          "current_content": { "type": "string" },
          "baseline_content": { "type": "string" }
        }
      },
...
    }
  }
}